Dispatch volume dashboard -including static zones
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@21146 842219ff-9341-0410-a943-573460bdd4de

+++ projects/Delivery/src/com/freshdirect/analytics/DispatchVolumeModel.java
@@ -8,7 +8,7 @@ public class DispatchVolumeModel {
 	private RoutingTimeOfDay dispatchTime;
 	private Date snapshotTime;
 	private Date dispatchDate;
-	private int plannedCapacity;
+	private double plannedCapacity;
 	private int plantCapacity;
 	private int orderCount;
 	private int noOftrucks;
@@ -18,10 +18,10 @@ public class DispatchVolumeModel {
 	public void setDispatchTime(RoutingTimeOfDay dispatchTime) {
 		this.dispatchTime = dispatchTime;
 	}
-	public int getPlannedCapacity() {
+	public double getPlannedCapacity() {
 		return plannedCapacity;
 	}
-	public void setPlannedCapacity(int plannedCapacity) {
+	public void setPlannedCapacity(double plannedCapacity) {
 		this.plannedCapacity = plannedCapacity;
 	}
 	public int getPlantCapacity() {
